+++ phoenix/grid.py
@@ -24,6 +24,9 @@ class MyGrid(Grid):
         mytemplate = mylookup.get_template(renderer + ".mako")
         return HTML.td(HTML.literal(mytemplate.render(**data)))
 
+    def render_label_td(self, label):
+        return self.render_td(renderer="label_td", label=label)
+
     def render_title_td(self, title, abstract="", keywords=[], data=[], format=None, source="#"):
         return self.render_td(renderer="title_td", title=title, abstract=abstract, keywords=keywords, data=data, format=format, source=source)
 
@@ -43,17 +46,17 @@ class MyGrid(Grid):
         return self.render_td(renderer="status_td", item=item, span_class=span_class)
     
     def render_timestamp_td(self, timestamp):
+        import datetime
         if timestamp is None:            
             return HTML.td('')
-        mytimestamp = timestamp
-        if type(timestamp) is str:
+        if type(timestamp) is not datetime.datetime:
             from dateutil import parser as datetime_parser
-            mytimestamp = datetime_parser.parse(timestamp)
+            timestamp = datetime_parser.parse(str(timestamp))
         span_class = 'due-date badge'
         
         span = HTML.tag(
             "span",
-            c=HTML.literal(mytimestamp.strftime('%Y-%m-%d %H:%M:%S')),
+            c=HTML.literal(timestamp.strftime('%Y-%m-%d %H:%M:%S')),
             class_=span_class,
         )
         return HTML.td(span)
@@ -144,7 +147,7 @@ class MyGrid(Grid):
         records.append(HTML.tag('thead', r))
         # now lets render the actual item grid
         for i, record in enumerate(self.itemlist):
-            logger.debug('item %s %s', i, record)
+            #logger.debug('item %s %s', i, record)
             columns = self.make_columns(i, record)
             if hasattr(self, 'custom_record_format'):
                 r = self.custom_record_format(i + 1, record, columns)