+++ phoenix/panels.py
@@ -1,5 +1,9 @@
 from pyramid_layout.panel import panel_config
 from pyramid.security import authenticated_userid, has_permission
+from pyramid.httpexceptions import HTTPException, HTTPFound, HTTPNotFound
+
+from deform import Form, ValidationFailure
+from phoenix import models
 
 import logging
 logger = logging.getLogger(__name__)
@@ -20,10 +24,10 @@ def navbar(context, request):
         nav.append( nav_item('Wizard', request.route_url('wizard')) )
         # TODO: enable map again when it is working
         #nav.append( nav_item('Map', request.route_url('map')) )
-        nav.append( nav_item('My Account', request.route_url('myaccount')) )
+        nav.append( nav_item('My Account', request.route_url('myaccount', tab='profile')) )
     if has_permission('admin', request.context, request):
         nav.append( nav_item('Settings', request.route_url('settings')) )
-    #nav.append( nav_item('Help', request.route_url('help')) )
+    nav.append( nav_item('Help', request.registry.settings.get('help.url')) )
 
     login = request.current_route_url() == request.route_url('signin', tab='esgf')
 
@@ -56,7 +60,7 @@ def logon_esgf(context, request):
 
 @panel_config(name='dashboard_users', renderer='phoenix:templates/panels/dashboard_users.pt')
 def dashboard_users(context, request):
-    from .models import user_stats
+    from phoenix.models import user_stats
     return user_stats(request)
 
 @panel_config(name='dashboard_jobs', renderer='phoenix:templates/panels/dashboard_jobs.pt')
@@ -65,6 +69,70 @@ def dashboard_jobs(context, request):
                 started = request.db.jobs.find({"is_complete": False}).count(),
                 failed = request.db.jobs.find({"is_complete": True, "is_succeded": False}).count(),
                 succeded = request.db.jobs.find({"is_succeded": True}).count())
+
+class MyAccountPanel(object):
+    def __init__(self, context, request):
+        self.context = context
+        self.request = request
+
+    def appstruct(self):
+        appstruct = models.get_user(self.request)
+        if appstruct is None:
+            appstruct = {}
+        return appstruct
+
+class MyAccoutProfile(MyAccountPanel):
+    def generate_form(self):
+        from phoenix.schema import UserProfileSchema
+        form = Form(schema=UserProfileSchema(), buttons=('update',), formid='deform')
+        return form
+    
+    def process_form(self, form):
+        try:
+            controls = self.request.POST.items()
+            appstruct = form.validate(controls)
+            user = models.get_user(self.request)
+            for key in ['name', 'organisation', 'notes']:
+                user[key] = appstruct.get(key)
+            self.request.db.users.update({'email':models.user_email(self.request)}, user)
+        except ValidationFailure, e:
+            logger.exception('validation of form failed.')
+            return dict(form=e.render())
+        except Exception, e:
+            logger.exception('update user failed.')
+            self.request.session.flash('Update of your accound failed. %s' % (e), queue='error')
+        else:
+            self.request.session.flash("Your account was updated.", queue='success')
+        #return HTTPFound(location=request.route_url('myaccount', tab='profile'))
+
+    @panel_config(name='myaccount_profile', renderer='phoenix:templates/panels/myaccount_default.pt')
+    def panel(self):
+        form = self.generate_form()
+        if 'update' in self.request.POST:
+            self.process_form(form)
+        return dict(title="Profile", form=form.render( self.appstruct() ))
+
+class MyAccountESGF(MyAccountPanel):
+    def generate_form(self):
+        from phoenix.schema import ESGFCredentialsSchema
+        form = Form(schema=ESGFCredentialsSchema(), formid='deform')
+        return form
+
+    @panel_config(name='myaccount_esgf', renderer='phoenix:templates/panels/myaccount_default.pt')
+    def panel(self):
+        form = self.generate_form()
+        return dict(title="ESGF Access", form=form.render( self.appstruct() ))
+
+class MyAccountSwift(MyAccountPanel):
+    def generate_form(self):
+        from phoenix.schema import SwiftSchema
+        form = Form(schema=SwiftSchema(), formid='deform')
+        return form
+
+    @panel_config(name='myaccount_swift', renderer='phoenix:templates/panels/myaccount_default.pt')
+    def panel(self):
+        form = self.generate_form()
+        return dict(title="Swift Access", form=form.render( self.appstruct() ))
 
 @panel_config(name='headings')
 def headings(context, request):