+++ phoenix/views/processes/executeprocess.py
@@ -35,7 +35,7 @@ class ExecuteProcess(Processes):
     def generate_form(self, formid='deform'):
         from phoenix.wps import WPSSchema
         # TODO: should be WPSSchema.bind() ...
-        schema = WPSSchema(info=True, process = self.process)
+        schema = WPSSchema(info=True, process = self.process, user=self.get_user())
         return Form(
             schema,
             buttons=('submit',),
@@ -56,7 +56,7 @@ class ExecuteProcess(Processes):
                 title = execution.process.title,
                 wps_url = execution.serviceInstance,
                 status_location = execution.statusLocation,
-                abstract = appstruct.get('abstract', ''),
+                abstract = execution.process.abstract,
                 keywords = appstruct.get('keywords', ''))
         except ValidationFailure, e:
             logger.exception('validation of exectue view failed.')